DXP-1568 Refactor attributes indexer
